+++ packages/box/src/component.tsx
@@ -6,26 +6,22 @@ import { box as ccBox } from '@warp-ds/css/component-classes';
 import { BoxProps } from './props.js';
 
 export function Box(props: BoxProps) {
-  const { children, as = 'div', bleed, clickable, neutral, bordered, info, role, ...rest } = props;
+  const { children, as = 'div', bleed, neutral, bordered, info, role, ...rest } = props;
+
+  const wrapperClasses = classNames(props.className, [
+    ccBox.base,
+    bleed && ccBox.bleed,
+    info && ccBox.info,
+    neutral && ccBox.neutral,
+    bordered && ccBox.bordered,
+  ]);
 
   return React.createElement(
     as,
     {
       ...(rest as Omit<BoxProps, 'children'> as {}),
       role: role ?? 'region',
-      className: classNames(
-        ccBox.box,
-        {
-          [ccBox.bleed]: bleed,
-          [ccBox.info]: info,
-          [ccBox.neutral]: neutral,
-          [ccBox.bordered]: bordered,
-          [ccBox.infoClickable]: clickable && info,
-          [ccBox.neutralClickable]: clickable && neutral,
-          [ccBox.borderedClickable]: clickable && bordered,
-        },
-        props.className,
-      ),
+      className: wrapperClasses,
     },
     children,
   );