+++ packages/switch/src/component.tsx
@@ -5,11 +5,31 @@ import { switchToggle as ccSwitch } from '@warp-ds/css/component-classes';
 
 import { SwitchProps } from './props.js';
 
-export function Switch({ id, value, onClick, 'aria-label': ariaLabel, 'aria-labelledby': ariaLabelledBy, ...attrs }: SwitchProps) {
-  const switchFocus = 'focusable rounded-full';
+export function Switch({
+  id,
+  value,
+  disabled,
+  onClick,
+  'aria-label': ariaLabel,
+  'aria-labelledby': ariaLabelledBy,
+  ...attrs
+}: SwitchProps) {
+  const baseClasses = classNames([ccSwitch.base, disabled && ccSwitch.disabled]);
+
+  const trackClasses = classNames([
+    ccSwitch.track,
+    disabled && ccSwitch.trackDisabled,
+    !disabled && (value ? ccSwitch.trackActive : ccSwitch.trackInactive),
+  ]);
+
+  const handleClasses = classNames([
+    ccSwitch.handle,
+    value && ccSwitch.handleSelected,
+    disabled ? ccSwitch.handleDisabled : ccSwitch.handleNotDisabled,
+  ]);
 
   return (
-    <div className={ccSwitch.switch}>
+    <div>
       <button
         id={id}
         type="button"
@@ -18,15 +38,12 @@ export function Switch({ id, value, onClick, 'aria-label': ariaLabel, 'aria-labe
         aria-labelledby={ariaLabelledBy}
         aria-checked={value}
         onClick={onClick}
-        className={classNames([ccSwitch.label, switchFocus])}
+        className={baseClasses}
+        aria-disabled={disabled}
+        disabled={disabled}
         {...attrs}>
-        <span data-testid="track" className={classNames([ccSwitch.track, value ? ccSwitch.trackActive : ccSwitch.trackInactive])} />
-        <span
-          data-testid="handle"
-          className={classNames([ccSwitch.handle, ccSwitch.handleNotDisabled], {
-            [ccSwitch.handleSelected]: value,
-          })}
-        />
+        <span data-testid="track" className={trackClasses} />
+        <span data-testid="handle" className={handleClasses} />
       </button>
     </div>
   );