+++ packages/textarea/src/component.tsx
@@ -2,11 +2,13 @@ import React, { forwardRef, useRef } from 'react';
 
 import { classNames } from '@chbphone55/classnames';
 import { i18n } from '@lingui/core';
-import { input as ccInput, label as ccLabel, helpText as ccHelpText } from '@warp-ds/css/component-classes';
+import { input as ccInput, label as ccLabel } from '@warp-ds/css/component-classes';
 
+import { HelpText } from '../../_helpers/help-text.js';
 import { activateI18n } from '../../i18n.js';
 import { useId } from '../../utils/src/index.js';
 
+import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
@@ -19,7 +21,6 @@ export const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>((props, f
   const {
     className,
     disabled,
-    error,
     helpText,
     id: providedId,
     invalid,
@@ -34,13 +35,23 @@ export const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>((props, f
     ...rest
   } = props;
 
-  activateI18n(enMessages, nbMessages, fiMessages);
+  activateI18n(enMessages, nbMessages, fiMessages, daMessages);
 
   const id = useId(providedId);
   const ref = useRef<HTMLTextAreaElement | null>(null);
 
   const helpId = helpText ? `${id}__hint` : undefined;
-  const isInvalid = invalid ?? error;
+  const isInvalid = invalid;
+
+  const textareaClasses = classNames([
+    ccInput.base,
+    ccInput.textArea,
+    !!placeholder && ccInput.placeholder,
+    !isInvalid && !disabled && !readOnly && ccInput.default,
+    isInvalid && !disabled && !readOnly && ccInput.invalid,
+    !isInvalid && disabled && !readOnly && ccInput.disabled,
+    !isInvalid && !disabled && readOnly && ccInput.readOnly,
+  ]);
 
   useTextAreaHeight({
     ref,
@@ -52,30 +63,21 @@ export const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>((props, f
   return (
     <div className={className} style={style}>
       {label && (
-        <label htmlFor={id} className={ccLabel.label}>
+        <label htmlFor={id} className={ccLabel.base}>
           {label}
           {optional && (
             <span className={ccLabel.optional}>
-              {i18n._(
-                /*i18n*/ {
-                  id: 'textarea.label.optional',
-                  message: '(optional)',
-                  comment: 'Shown behind label when marked as optional',
-                },
-              )}
+              {i18n._({
+                id: 'textarea.label.optional',
+                message: '(optional)',
+                comment: 'Shown behind label when marked as optional',
+              })}
             </span>
           )}
         </label>
       )}
       <textarea
-        className={classNames({
-          [`${ccInput.base} ${ccInput.textArea}`]: true,
-          [ccInput.placeholder]: !!placeholder,
-          [ccInput.default]: !isInvalid && !disabled && !readOnly,
-          [ccInput.invalid]: isInvalid && !disabled && !readOnly,
-          [ccInput.disabled]: !isInvalid && disabled && !readOnly,
-          [ccInput.readOnly]: !isInvalid && !disabled && readOnly,
-        })}
+        className={textareaClasses}
         {...rest}
         placeholder={placeholder}
         aria-describedby={helpId}
@@ -97,17 +99,7 @@ export const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>((props, f
         readOnly={readOnly}
         value={value}
       />
-      {helpText && (
-        <div
-          className={classNames({
-            [ccHelpText.helpText]: true,
-            [ccHelpText.helpTextColor]: !isInvalid,
-            [ccHelpText.helpTextColorInvalid]: isInvalid,
-          })}
-          id={helpId}>
-          {helpText}
-        </div>
-      )}
+      {props.helpText && <HelpText helpId={helpId} helpText={helpText} isInvalid={isInvalid} />}
     </div>
   );
 });