+++ packages/textfield/src/component.tsx
@@ -2,11 +2,13 @@ import React, { forwardRef } from 'react';
 
 import { classNames } from '@chbphone55/classnames';
 import { i18n } from '@lingui/core';
-import { input as ccInput, label as ccLabel, helpText as ccHelpText } from '@warp-ds/css/component-classes';
+import { input as ccInput, label as ccLabel } from '@warp-ds/css/component-classes';
 
+import { HelpText } from '../../_helpers/help-text.js';
 import { activateI18n } from '../../i18n.js';
 import { useId } from '../../utils/src/index.js';
 
+import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
@@ -19,7 +21,6 @@ export const TextField = forwardRef<HTMLInputElement, TextFieldProps>((props, re
     id: providedId,
     children,
     invalid,
-    error,
     helpText,
     label,
     readOnly,
@@ -29,29 +30,38 @@ export const TextField = forwardRef<HTMLInputElement, TextFieldProps>((props, re
     ...rest
   } = props;
 
-  activateI18n(enMessages, nbMessages, fiMessages);
+  activateI18n(enMessages, nbMessages, fiMessages, daMessages);
 
   const id = useId(providedId);
   const helpId = helpText ? `${id}__hint` : undefined;
-  const isInvalid = invalid || error;
+  const isInvalid = invalid;
 
   const suffix = React.Children.toArray(children).find((child) => React.isValidElement(child) && child.props.suffix);
   const prefix = React.Children.toArray(children).find((child) => React.isValidElement(child) && child.props.prefix);
 
+  const inputClasses = classNames([
+    ccInput.base,
+    !!props.placeholder && ccInput.placeholder,
+    !!suffix && ccInput.suffix,
+    !!prefix && ccInput.prefix,
+    !isInvalid && !disabled && !readOnly && ccInput.default,
+    isInvalid && !disabled && !readOnly && ccInput.invalid,
+    !isInvalid && disabled && !readOnly && ccInput.disabled,
+    !isInvalid && !disabled && readOnly && ccInput.readOnly,
+  ]);
+
   return (
     <div className={className} style={style}>
       {label && (
-        <label htmlFor={id} className={ccLabel.label}>
+        <label htmlFor={id} className={ccLabel.base}>
           {label}
           {optional && (
             <span className={ccLabel.optional}>
-              {i18n._(
-                /*i18n*/ {
-                  id: 'textfield.label.optional',
-                  message: '(optional)',
-                  comment: 'Shown behind label when marked as optional',
-                },
-              )}
+              {i18n._({
+                id: 'textfield.label.optional',
+                message: '(optional)',
+                comment: 'Shown behind label when marked as optional',
+              })}
             </span>
           )}
         </label>
@@ -59,16 +69,7 @@ export const TextField = forwardRef<HTMLInputElement, TextFieldProps>((props, re
       <div className={ccInput.wrapper}>
         {prefix}
         <input
-          className={classNames({
-            [ccInput.base]: true,
-            [ccInput.default]: !isInvalid && !disabled && !readOnly,
-            [ccInput.invalid]: isInvalid && !disabled && !readOnly,
-            [ccInput.disabled]: !isInvalid && disabled && !readOnly,
-            [ccInput.readOnly]: !isInvalid && !disabled && readOnly,
-            [ccInput.placeholder]: !!props.placeholder,
-            [ccInput.suffix]: !!suffix,
-            [ccInput.prefix]: !!prefix,
-          })}
+          className={inputClasses}
           {...rest}
           aria-describedby={helpId}
           aria-errormessage={isInvalid && helpId ? helpId : undefined}
@@ -82,17 +83,7 @@ export const TextField = forwardRef<HTMLInputElement, TextFieldProps>((props, re
         {suffix}
       </div>
 
-      {helpText && (
-        <div
-          className={classNames({
-            [ccHelpText.helpText]: true,
-            [ccHelpText.helpTextColor]: !isInvalid,
-            [ccHelpText.helpTextColorInvalid]: isInvalid,
-          })}
-          id={helpId}>
-          {helpText}
-        </div>
-      )}
+      {props.helpText && <HelpText helpId={helpId} helpText={helpText} isInvalid={isInvalid} />}
     </div>
   );
 });