+++ tests/components/CardTest.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { fireEvent, render, screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 
 import { Card } from '../../packages/card/src/component';
 
@@ -39,31 +39,5 @@ describe('Card', () => {
     expect(screen.getByRole('article')).toHaveClass(
       'cursor-pointer overflow-hidden relative transition-all border-2 rounded-4 s-bg-selected hover:s-bg-selected-hover active:s-bg-selected-active s-border-selected hover:s-border-selected-hover active:s-border-selected-active',
     );
-  });
-
-  it('handles onKeyDown events', () => {
-    const handleClick = vi.fn();
-    render(<Card as="article" onClick={handleClick} />);
-    fireEvent.keyDown(screen.getByRole('article'), { key: 'Enter', code: 'Enter' });
-    expect(handleClick).toHaveBeenCalledTimes(1);
-    fireEvent.keyDown(screen.getByRole('article'), { key: ' ', code: 'Space' });
-    expect(handleClick).toHaveBeenCalledTimes(2);
-  });
-
-  it('does not call onClick for other keys', () => {
-    const handleClick = vi.fn();
-    render(<Card as="article" onClick={handleClick} />);
-    fireEvent.keyDown(screen.getByRole('article'), { key: 'ArrowRight', code: 'ArrowRight' });
-    expect(handleClick).not.toHaveBeenCalled();
-  });
-
-  it('renders sr-only button when onClick is provided', () => {
-    render(<Card as="article" onClick={() => {}} />);
-    expect(screen.getByRole('button')).toBeInTheDocument();
-  });
-
-  it('does not render sr-only button when onClick is not provided', () => {
-    render(<Card />);
-    expect(screen.queryByRole('button')).toBeNull();
   });
 });